+++ docs/examples/UserGuide/misc_tips.jl
@@ -13,8 +13,8 @@ dfv = dt(db, df, "dfv");
 
 # ## DuckDB's SUMMARIZE
 # DuckDB has a feature tosummarize tables that gives information about the table, such as mean, std, q25, q75 etc.
-# To use this feature with TidierDB, simply call an empty `@summarize`. 
-@chain dfv @summarize() @collect
+# To use this feature with TidierDB, simply call an `@summary` on any table or file before querying it.
+@chain dfv @summary() @collect
 
 # ## show_query/collect
 # If you find yourself frequently showing a query while collecting, you can define the following function 
@@ -25,7 +25,7 @@ sqc(qry) = @chain qry begin
 
 # Call this function at the end of a chain similar the `@show_query` or`@collect` macros
 # _printed query is not seen here as it prints to the REPL_
-@chain dfv @summarize() sqc()
+@chain dfv @summary() sqc()
 
 # ## Color Printing
 # Queries print with some code words in color to the REPL. To turn off this feature, run one of the following.